requirements.txt
beautifulsoup4
discord.py
lxml
markdownify
py-stackexchange
PyYAML
six
python-dateutil
psutil
asyncpg